Avoid calling +hash: on the main thread by using two keys in the map table for each image (URL string and hash).
Calls to +hash: on the main thread were leading to a noticeable imprint on older hardware in instruments. To avoid calling +hash: on main, which was only needed to check our map table and delegate map, we switch the map table to having two entries per image: one with the URL string as the key and the other with the hash of the URL string as the key. Now the map table can be checked using either a URL string as the key -or- using the hash as the key.