Clean up files on failure.
There's evidence the system doesn't automatically clean these up in this case even though the API claims to.

https://dropbox.com/s/mxzqvs402o3rx7e

Perhaps it has something to do with how I inspect the results? In any case, since these are in tmp they'll be cleaned up at some point, but why not do it eagerly?